% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphi.r
\name{delphi}
\alias{delphi}
\title{Function to compute derivative of likelihood with respect to the dispersion parameter \eqn{\phi}}
\usage{
delphi(y = NULL, xi = NULL, mu = NULL, phi = NULL, h = 1e-10)
}
\arguments{
\item{y}{observed response}

\item{xi}{index parameter}

\item{mu}{mean vector}

\item{phi}{dispersion vector}
}
\description{
Internal Function
}
