% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.tw.R
\name{dev.tw}
\alias{dev.tw}
\title{Deviance function for Tweedie Compound Poisson-gamma distribution}
\usage{
dev.tw(y = NULL, wts = NULL, mu = NULL, xi = NULL)
}
\arguments{
\item{y}{observed response}

\item{wts}{optional weights (or scales) for y}

\item{mu}{mean vector}

\item{xi}{index parameter}
}
\description{
Internal Function
}
