% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dglm.autograd.R
\name{dglm.autograd}
\alias{dglm.autograd}
\title{An MCMC sampler for fitting Tweedie Compound Poisson-Gamma (CP-g) Double Generalized Linear Models}
\usage{
dglm.autograd(
  y = NULL,
  x = NULL,
  z = NULL,
  beta.init = NULL,
  gamma.init = NULL,
  xi.init = NULL,
  prec.beta = NULL,
  prec.gamma = NULL,
  lower.xi = NULL,
  upper.xi = NULL,
  tau.beta = 0.1,
  tau.gamma = 0.1,
  tau.xi = 0.1,
  niter = NULL,
  nburn = NULL,
  report = NULL,
  thin = 1,
  return.mcmc = T,
  verbose = F,
  digits = 3
)
}
\arguments{
\item{y}{observed response}

\item{x}{covariates for the mean model}

\item{z}{covariates for the dispersion model}

\item{lower.xi}{lower bound for the index parameter of the Tweedie CP-g distribution}

\item{upper.xi}{upper bound for the index parameter of the Tweedie CP-g distribution}
}
\description{
Fits a univariate Gaussian spatial regression model: \eqn{\log(\mu)=x^T\beta} and \eqn{\log(\phi)=z^T\gamma}. Parameters not listed below are optional.
}
