% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR.R
\name{FDR}
\alias{FDR}
\title{Computes False discovery rate using Bayesian q-values}
\usage{
FDR(chain = NULL, thres = 0.05, FDRc = 0.05)
}
\arguments{
\item{chain}{MCMC chain for beta or gamma in DGLM}

\item{thres}{threshold value}

\item{FDRc}{percent at which FDR is computed}
}
\description{
Computes False discovery rate using Bayesian q-values
}
\examples{
\dontrun{
rbeta.ssdglm = FDR(mc$beta.mcmc)
rgamma.ssdglm = FDR(mc$gamma.mcmc)
}
}
\keyword{FDR}
