% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mcmc.R
\name{plot_mcmc}
\alias{plot_mcmc}
\title{MCMC Diagnostic Plots for samplers}
\usage{
plot_mcmc(
  samples = NULL,
  true = NULL,
  col = "darkgreen",
  acf.lag.max = 100,
  cnames = NULL
)
}
\arguments{
\item{samples}{MCMC samples}

\item{true}{true values, optional}

\item{cnames}{parameter names}
}
\description{
Internal Function
}
